adds additional support for integration models (#56)
This commit adds additional command support for integration models.  It
includes the following new commands:

- `import integration-model ...`
- `create integration-model ...`
- `clear integration-model ...`

+++ internal/runners/importer.go
@@ -22,11 +22,16 @@ import (
 func importUnmarshalFromRequest(in Request, ptr any) error {
 	logger.Trace()
 
+	common := in.Common.(*flags.AssetImportCommon)
+
 	path, err := importGetPathFromRequest(in)
 	if err != nil {
 		return err
 	}
-	defer os.RemoveAll(path)
+
+	if common.Repository != "" {
+		defer os.RemoveAll(path)
+	}
 
 	return importLoadFromDisk(path, ptr)
 }
